+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Ecosystem/Proof of Passport doc;
+- Passport-derived profiles overview doc;
 
 ## [3.4.1] - 25-04-2024