+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Passport-derived profiles overview doc;
+- Passport related terms to the glossary;
 
 ## [3.4.1] - 25-04-2024